requirements.txt
absl-py==0.9.0
anytree==2.8.0
appdirs==1.4.3
appnope==0.1.0
asgiref==3.2.7
asn1crypto==1.3.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.2
bleach==3.1.0
bootstrap4==0.1.0
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
conda==4.8.3
conda-build==3.18.11
conda-package-handling==1.6.0
conda-verify==3.4.2
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-chartjs==2.0.0
django-heroku==0.3.1
django-toolbelt==0.0.1
djangorestframework==3.11.0
entrypoints==0.3
filelock==3.0.12
future==0.18.2
gast==0.2.2
geocoder==1.38.1
glob2==0.7
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googlemaps==4.4.0
grpcio==1.28.1
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.1.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
libarchive-c==2.8
lief==0.9.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mysqlclient==1.4.6
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.2
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
opt-einsum==3.2.0
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pbr==5.4.4
pdf2image==1.13.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pkginfo==1.5.0.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.3
psutil==5.7.0
psycopg2==2.8.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycosat==0.6.3
pycparser==2.20
pygeocoder==1.2.5
Pygments==2.6.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
pyzmq==18.1.1
QtPy==1.9.0
ratelim==0.1.6
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel-yaml==0.15.87
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
static3==0.7.0
stevedore==1.32.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm==4.44.1
traitlets==4.3.3
urllib3==1.25.8
virtualenv==20.0.15
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
